requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
certifi==2020.4.5.1
chardet==3.0.4
idna==2.9
idna-ssl==1.1.0
multidict==4.7.5
pkg-resources==0.0.0
pysmartthings==0.7.1
python-crontab==2.4.1
python-dateutil==2.8.1
pytz==2019.3
requests==2.23.0
six==1.14.0
typing-extensions==3.7.4.2
urllib3==1.25.8
yarl==1.4.2
